Metastore: to use common scripts infrastructure

+++ python/cloudtik/runtime/metastore/scripts/services.sh
@@ -5,24 +5,42 @@ if [ ! -n "${METASTORE_HOME}" ]; then
     exit 1
 fi
 
-case "$1" in
-start-head)
-    if [ "${CLOUD_DATABASE}" != "true" ] || [ "$METASTORE_WITH_CLOUD_DATABASE" == "false" ]; then
-        sudo service mysql start
+command=$1
+shift
+
+# Parsing arguments
+IS_HEAD_NODE=false
+
+while [[ $# -gt 0 ]]
+do
+    key="$1"
+    case $key in
+    -h|--head)
+        IS_HEAD_NODE=true
+        ;;
+    *)
+        echo "Unknown argument passed."
+        exit 1
+    esac
+    shift
+done
+
+case "$command" in
+start)
+    if [ $IS_HEAD_NODE == "true" ]; then
+        if [ "${CLOUD_DATABASE}" != "true" ] || [ "$METASTORE_WITH_CLOUD_DATABASE" == "false" ]; then
+            sudo service mysql start
+        fi
+        nohup $METASTORE_HOME/bin/start-metastore >${METASTORE_HOME}/logs/start-metastore.log 2>&1 &
     fi
-    nohup $METASTORE_HOME/bin/start-metastore >${METASTORE_HOME}/logs/start-metastore.log 2>&1 &
     ;;
-stop-head)
-    if [ "${CLOUD_DATABASE}" != "true" ] || [ "$METASTORE_WITH_CLOUD_DATABASE" == "false" ]; then
-        sudo service mysql stop
+stop)
+    if [ $IS_HEAD_NODE == "true" ]; then
+        if [ "${CLOUD_DATABASE}" != "true" ] || [ "$METASTORE_WITH_CLOUD_DATABASE" == "false" ]; then
+            sudo service mysql stop
+        fi
+        pkill -f 'HiveMetaStore'
     fi
-    pkill -f 'HiveMetaStore'
-    ;;
-start-worker)
-    # No need to run anything for worker node
-    ;;
-stop-worker)
-    # No need to run anything for worker node
     ;;
 -h|--help)
     echo "Usage: $0 start|stop --head" >&2
